refactor: remove unnecessary AuthenticationManager usage in JwtFilter
