+++ src/main/java/com/habittracker/api/security/jwt/filter/JwtFilter.java
@@ -12,11 +12,6 @@ import java.util.Optional;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -30,7 +25,6 @@ public class JwtFilter extends OncePerRequestFilter {
 
   private final JwtService jwtService;
   private final UserDetailsService userDetailsService;
-  private final AuthenticationManager authenticationManager;
 
   @Override
   protected void doFilterInternal(
@@ -43,7 +37,7 @@ public class JwtFilter extends OncePerRequestFilter {
         .flatMap(jwtService::extractSubject)
         .flatMap(this::subjectToUserDetails)
         .map(SecurityUtils::userDetailsToAuthenticationToken)
-        .ifPresent(this::populateSecurityContext);
+        .ifPresent(SecurityUtils::populateSecurityContext);
     filterChain.doFilter(request, response);
   }
 
@@ -55,14 +49,4 @@ public class JwtFilter extends OncePerRequestFilter {
     }
   }
 
-  private void populateSecurityContext(UsernamePasswordAuthenticationToken token)
-      throws AuthenticationException {
-    try {
-      Authentication authenticate = authenticationManager.authenticate(token);
-      SecurityContextHolder.getContext().setAuthentication(authenticate);
-    } catch (AuthenticationException ex) {
-      log.warn("Authentication failed: invalid credentials");
-      log.debug("Authentication error details: {}", ex.getMessage());
-    }
-  }
 }
